Refactoring & MultiThreading
